+++ src/components/data/EntitiesContent.js
@@ -10,7 +10,7 @@ import { tools } from 'src/analysis/utils/tool-utils';
 import { requesterPaysWrapper, withRequesterPaysHandler } from 'src/components/bucket-utils';
 import { ButtonSecondary } from 'src/components/common';
 import { DataTableColumnProvenance } from 'src/components/data/data-table-provenance';
-import { AddColumnModal, CreateEntitySetModal, ModalToolButton } from 'src/components/data/data-utils';
+import { CreateEntitySetModal, ModalToolButton } from 'src/components/data/data-utils';
 import DataTable from 'src/components/data/DataTable';
 import ExportDataModal from 'src/components/data/ExportDataModal';
 import { icon, spinner } from 'src/components/icons';
@@ -22,6 +22,7 @@ import { withModalDrawer } from 'src/components/ModalDrawer';
 import { MenuDivider, MenuTrigger } from 'src/components/PopupTrigger';
 import TitleBar from 'src/components/TitleBar';
 import WorkflowSelector from 'src/components/WorkflowSelector';
+import { AddColumnModal } from 'src/data/data-table/entity-service/AddColumnModal';
 import { AddEntityModal } from 'src/data/data-table/entity-service/AddEntityModal';
 import { entityAttributeText } from 'src/data/data-table/entity-service/entityAttributeText';
 import { EntityDeleter } from 'src/data/data-table/entity-service/EntityDeleter';
